readme, comments, refresh, modularization
