+++ README.md
@@ -1,1 +1,161 @@
-# movies-capstone-project
+<div align="center">
+  <h3><b>Your Amazing Shows</b></h3>
+
+</div>
+# 📗 Table of Contents
+
+- [📗 Table of Contents](#-table-of-contents)
+- [📖 JavaScript Capstone](#-to-do-list-)
+  - [🛠 Built With](#-built-with-)
+    - [Tech Stack](#tech-stack-)
+    - [Key Features](#key-features-)
+    - [🚀 Live Demo](#live-demo)
+  - [💻 Getting Started](#-getting-started-)
+    - [Prerequisites](#prerequisites)
+    - [Setup](#setup)
+    - [Install](#install)
+    - [Usage](#usage)
+    - [Run tests](#run-tests)
+    - [Deployment](#deployment)
+  - [👥 Authors](#-authors-)
+  - [🔭 Future Features](#-future-features-)
+  - [🤝 Contributing](#-contributing-)
+  - [⭐️ Show your support](#️-show-your-support-)
+  - [📝 License](#-license-)
+
+# 📖 JavaScript Capstone <a name="about-project"></a>
+
+**Your Amazing Shows** The project delivers a simple web project in HTML, CSS, and JavaScript that consumes a movies api and microverse api supporting, commenting and liking.
+
+## 🛠 Built With <a name="built-with"></a>
+
+### Tech Stack <a name="tech-stack"></a>
+
+<details>
+  <summary>Client</summary>
+  <ul>
+    <li>HTML</li>
+    <li>Javascript</li>
+    <li>CSS</li>
+  </ul>
+</details>
+
+### Key Features <a name="key-features"></a>
+
+- **A landing page showing different movies you can like"**
+- **A popup window with more details about the show**
+- **Additionally in the popup, they can see comments and add comments**
+
+<p align="right">(<a href="#readme-top">back to top</a>)</p>
+
+## 🚀 Live Demo <a name="live-demo"></a>
+
+-[Live Demo Link](https://king-benx.github.io/your-movies-capstone-project/)
+
+## 🚀 Project documentation <a name="documentation"></a>
+
+-[Walkthrough]()
+
+<p align="right">(<a href="#readme-top">back to top</a>)</p>
+
+## 💻 Getting Started <a name="getting-started"></a>
+
+To get a local copy up and running, follow these steps.
+
+### Prerequisites
+
+In order to run this project you need:
+
+- A browser
+- Node v16
+- Git
+
+### Setup
+
+Clone this repository to your desired folder:
+
+```sh
+  cd your-desired-folder
+  git clone git@github.com:King-Benx/your-movies-capstone-project.git
+```
+
+### Install
+
+Install this project dependencies with:
+
+```sh
+  npm install
+```
+
+### Usage
+
+To run the project, execute the following command:
+
+- open the repo folder `your-movies-capstone-project` and run the following command within your terminal
+
+```sh
+ npm run dev
+```
+
+### Run tests
+
+To run tests to check the lint errors, run the following command:
+
+```sh
+  npx eslint . --fix
+```
+
+To run unit tests, run the following command:
+
+```sh
+  npm test
+```
+
+### Deployment
+
+You can deploy this project using:
+
+```sh
+  npm run predeploy && npm run deploy
+
+```
+
+<p align="right">(<a href="#readme-top">back to top</a>)</p>
+
+## 👥 Authors <a name="authors"></a>
+
+👤 **Author1**
+
+- GitHub: [Benard Asiimwe Bataringaya](https://github.com/King-Benx)
+
+👤 **Author2**
+
+- GitHub: [Dilsher Balouch](https://github.com/DilsherB)
+
+<p align="right">(<a href="#readme-top">back to top</a>)</p>
+
+## 🔭 Future Features <a name="future-features"></a>
+
+- [ ] Tests.
+
+<p align="right">(<a href="#readme-top">back to top</a>)</p>
+
+## 🤝 Contributing <a name="contributing"></a>
+
+Contributions, issues, and feature requests are welcome!
+
+Feel free to check the [issues page](https://github.com/King-Benx/your-movies-capstone-project/issues).
+
+<p align="right">(<a href="#readme-top">back to top</a>)</p>
+
+## ⭐️ Show your support <a name="support"></a>
+
+If you like this project...
+
+<p align="right">(<a href="#readme-top">back to top</a>)</p>
+
+## 📝 License <a name="license"></a>
+
+This project is [MIT](./LICENSE) licensed.
+
+<p align="right">(<a href="#readme-top">back to top</a>)</p>
