+++ src/index.html
@@ -34,7 +34,7 @@
             <tbody id="table-body"></tbody>
           </table>
           <h4>Add a comment</h4>
-          <form>
+          <form id="form">
             <div class="form-group">
               <input
                 type="text"
@@ -48,15 +48,13 @@
               <textarea
                 name="comment"
                 id="comment"
-                cols="30"
-                rows="10"
                 required
                 placeholder="Your insights"
               >
               </textarea>
             </div>
             <input type="hidden" id="item-id">
-            <button type="submit">Comment</button>
+            <button type="submit" id="submit">Comment</button>
           </form>
         </div>
       </dialog>
