+++ src/styles/style.scss
@@ -79,7 +79,9 @@ dialog {
       }
 
       textarea {
+        text-align: left;
         width: 100%;
+        height: 80px;
       }
     }
